+++ packages/lumx-react/src/components/select/Select.tsx
@@ -13,6 +13,7 @@ import { InputLabel } from '@lumx/react/components/input-label/InputLabel';
 
 import { Comp } from '@lumx/react/utils/type';
 import { getRootClassName, handleBasicClasses } from '@lumx/react/utils/className';
+import { mergeRefs } from '@lumx/react/utils/mergeRefs';
 
 import { WithSelectContext } from './WithSelectContext';
 import { CoreSelectProps, SelectVariant } from './constants';
@@ -61,6 +62,7 @@ const SelectField: React.FC<SelectProps> = ({
     theme,
     value,
     variant,
+    selectElementRef,
     ...forwardedProps
 }) => {
     return (
@@ -82,7 +84,7 @@ const SelectField: React.FC<SelectProps> = ({
 
                     {/* eslint-disable-next-line jsx-a11y/no-static-element-interactions */}
                     <div
-                        ref={anchorRef as RefObject<HTMLDivElement>}
+                        ref={mergeRefs(anchorRef as RefObject<HTMLDivElement>, selectElementRef)}
                         id={id}
                         className={`${CLASSNAME}__wrapper`}
                         onClick={onInputClick}
@@ -145,7 +147,7 @@ const SelectField: React.FC<SelectProps> = ({
                     after={<Icon icon={isEmpty ? mdiMenuDown : mdiCloseCircle} />}
                     onAfterClick={isEmpty ? onInputClick : onClear}
                     onClick={onInputClick}
-                    ref={anchorRef as RefObject<HTMLAnchorElement>}
+                    ref={mergeRefs(anchorRef as RefObject<HTMLAnchorElement>, selectElementRef)}
                     theme={theme}
                     {...forwardedProps}
                 >